Begin refactor to allow args
Committing to avoid lost work.

This might not be merged in. In order to properly support this, it has
become increasingly more likely that how interpolation is done will need
a complete overhaul to interpolate on a task-by-task basis rather than
doing the yaml file as a whole. This will probably have a number of side
effects with regards to how multi-line interpolation is handled as well.

Note that the changes to this branch should be reviewed critically if
restarted, as none of them have been tested.